+++ example/main.dart
@@ -11,7 +11,15 @@ class State {
 class Increment extends Action<State> {
   Increment([this.by = 1]);
   final int by;
-  State reduce(State state) => State(state.count + by);
+  State reduce(state) => State(state.count + by);
+
+  @override
+  void afterReduce(Store<State> store, State state) {
+    if (state.count < 10) {
+      store.dispatch(Increment(by + 1));
+    }
+    super.afterReduce(store, state);
+  }
 }
 
 void main() {
@@ -21,12 +29,5 @@ void main() {
   print(store.state.count); // 0
 
   store.dispatch(Increment());
-  // Before action: Increment: 0 (from LogMiddleware)
-  // After action: Increment: 1 (from LogMiddleware)
   print(store.state.count); // 1
-
-  store.dispatch(Increment(2));
-  // Before action: Increment: 1 (from LogMiddleware)
-  // After action: Increment: 3 (from LogMiddleware)
-  print(store.state.count); // 3
 }