+++ lib/src/log_middleware.dart
@@ -4,15 +4,13 @@ import '../redurx.dart';
 class LogMiddleware<T> extends Middleware<T> {
   /// Prints before the Action reducer call.
   @override
-  T beforeAction(store, action, state) {
-    print('Before action: ${action.runtimeType}: $state');
-    return super.beforeAction(store, action, state);
+  void beforeAction(store, action, state) {
+    print('MIDDLEWARE before: ${action.runtimeType}: $state');
   }
 
   /// Prints after the Action reducer call.
   @override
-  T afterAction(store, action, state) {
-    print('After action: ${action.runtimeType}: $state');
-    return super.afterAction(store, action, state);
+  void afterAction(store, action, state) {
+    print('MIDDLEWARE after: ${action.runtimeType}: $state');
   }
 }