+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+* **Breaking** Middleware does not modify the state
+* Action and AsyncAction can override afterReducer Method that has access to the store
+* AsyncAction can override shouldReduce, to decide if reduced has to be called
+
 # 1.1.0
 * **Breaking** Middlewares knows about the Store